+++ example.py
@@ -1,3 +1,5 @@
+import logging
+import sys
 from czbenchmarks.datasets.utils import load_dataset
 from czbenchmarks.runner import run_inference
 from czbenchmarks.tasks import (
@@ -5,28 +7,26 @@ from czbenchmarks.tasks import (
     EmbeddingTask,
     MetadataLabelPredictionTask,
 )
-from czbenchmarks.utils import get_aws_credentials
 
 if __name__ == "__main__":
-    aws_credentials = get_aws_credentials(profile="default")
-
-    dataset = load_dataset("example", config_path="custom.yaml")
+    logging.basicConfig(level=logging.INFO, stream=sys.stdout)
+    dataset = load_dataset("tsv2_bladder")
 
     for model_name in ["SCVI", "SCGPT"]:
-        dataset = run_inference(model_name, dataset, environment=aws_credentials)
+        dataset = run_inference(model_name, dataset)
 
-    task = ClusteringTask(label_key="cell_type")
-    clustering_results = task.run(dataset)
+        task = ClusteringTask(label_key="cell_type")
+        clustering_results = task.run(dataset)
 
-    task = EmbeddingTask(label_key="cell_type")
-    embedding_results = task.run(dataset)
+        task = EmbeddingTask(label_key="cell_type")
+        embedding_results = task.run(dataset)
 
-    task = MetadataLabelPredictionTask(label_key="cell_type")
-    prediction_results = task.run(dataset)
+        task = MetadataLabelPredictionTask(label_key="cell_type")
+        prediction_results = task.run(dataset)
 
-    print("Clustering results:")
-    print(clustering_results)
-    print("Embedding results:")
-    print(embedding_results)
-    print("Prediction results:")
-    print(prediction_results)
+        print("Clustering results:")
+        print(clustering_results)
+        print("Embedding results:")
+        print(embedding_results)
+        print("Prediction results:")
+        print(prediction_results)